+++ plugins/input/kafka/input_kafka.go
@@ -1,4 +1,4 @@
-// Copyright 2021 iLogtail Authors
+// Copyright 2023 iLogtail Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,15 +15,23 @@
 package kafka
 
 import (
+	ctx "context"
 	"fmt"
 	"strings"
 	"sync"
 
+	"github.com/Shopify/sarama"
+
 	"github.com/alibaba/ilogtail/pkg/logger"
 	"github.com/alibaba/ilogtail/pkg/pipeline"
+	"github.com/alibaba/ilogtail/pkg/pipeline/extensions"
+	"github.com/alibaba/ilogtail/pkg/protocol/decoder"
+	"github.com/alibaba/ilogtail/pkg/protocol/decoder/common"
+)
 
-	"github.com/Shopify/sarama"
-	cluster "github.com/bsm/sarama-cluster"
+const (
+	v1 = iota
+	v2
 )
 
 type InputKafka struct {
@@ -36,17 +44,29 @@ type InputKafka struct {
 	Offset        string
 	SASLUsername  string
 	SASLPassword  string
+	// Assignor Consumer group partition assignment strategy (range, roundrobin, sticky)
+	Assignor string
+	// Decoder the decoder to use, default is "ext_default_decoder"
+	Decoder           string
+	Format            string
+	FieldsExtend      bool
+	DisableUncompress bool
 
-	cluster  *cluster.Consumer
-	wg       *sync.WaitGroup
-	shutdown chan struct{}
-
-	context pipeline.Context
+	ready               chan bool
+	readyCloser         sync.Once
+	consumerGroupClient sarama.ConsumerGroup
+	wg                  *sync.WaitGroup
+	messages            chan *sarama.ConsumerMessage
+	context             pipeline.Context
+	cancelConsumer      ctx.CancelFunc
+	collectorV2         pipeline.PipelineCollector
+	decoder             extensions.Decoder
+	collectorV1         pipeline.Collector
+	version             int8
 }
 
 const (
-	maxInputKafkaLen = 512 * 1024
-	pluginName       = "service_kafka"
+	pluginName = "service_kafka"
 )
 
 func (k *InputKafka) Init(context pipeline.Context) (int, error) {
@@ -64,22 +84,27 @@ func (k *InputKafka) Init(context pipeline.Context) (int, error) {
 	if k.ClientID == "" {
 		return 0, fmt.Errorf("must specify ClientID for plugin %v", pluginName)
 	}
-	if k.MaxMessageLen > maxInputKafkaLen || k.MaxMessageLen < 1 {
-		return 0, fmt.Errorf("MaxMessageLen must be between 1 and %v for plugin %v",
-			maxInputKafkaLen, pluginName)
+
+	// init decoder
+	if k.Format == "" {
+		k.Format = common.ProtocolRaw
+	}
+	var err error
+	if k.decoder, err = decoder.GetDecoderWithOptions(k.Format, decoder.Option{
+		FieldsExtend:      k.FieldsExtend,
+		DisableUncompress: k.DisableUncompress,
+	}); err != nil {
+		return 0, err
 	}
 
-	config := cluster.NewConfig()
+	config := sarama.NewConfig()
 
 	if k.Version != "" {
-		version, err := sarama.ParseKafkaVersion(k.Version)
-		if err != nil {
+		if config.Version, err = sarama.ParseKafkaVersion(k.Version); err != nil {
 			return 0, err
 		}
-		config.Version = version
 	}
 	config.Consumer.Return.Errors = true
-	config.Group.Return.Notifications = true
 
 	if k.SASLUsername != "" && k.SASLPassword != "" {
 		logger.Infof(k.context.GetRuntimeContext(), "Using SASL auth with username '%s',",
@@ -100,19 +125,52 @@ func (k *InputKafka) Init(context pipeline.Context) (int, error) {
 		config.Consumer.Offsets.Initial = sarama.OffsetOldest
 	}
 
-	var clusterErr error
-	k.cluster, clusterErr = cluster.NewConsumer(
-		k.Brokers,
-		k.ConsumerGroup,
-		k.Topics,
-		config,
-	)
-	if clusterErr != nil {
-		err := fmt.Errorf("Error when creating Kafka consumer, brokers: %v, topics: %v, err: %v",
-			k.Brokers, k.Topics, clusterErr)
-		logger.Errorf(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM", "%v", err)
+	switch strings.ToLower(k.Assignor) {
+	case "sticky":
+		config.Consumer.Group.Rebalance.GroupStrategies = []sarama.BalanceStrategy{sarama.BalanceStrategySticky}
+	case "roundrobin":
+		config.Consumer.Group.Rebalance.GroupStrategies = []sarama.BalanceStrategy{sarama.BalanceStrategyRoundRobin}
+	case "range":
+		config.Consumer.Group.Rebalance.GroupStrategies = []sarama.BalanceStrategy{sarama.BalanceStrategyRange}
+	default:
+		logger.Warningf(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM", "Unrecognized consumer group partition assignor '%s', using 'oldest'",
+			k.Assignor)
+		config.Consumer.Group.Rebalance.GroupStrategies = []sarama.BalanceStrategy{sarama.BalanceStrategyRange}
+	}
+
+	newClient, err := sarama.NewClient(k.Brokers, config)
+	if err != nil {
+		logger.Warningf(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM", "Kafka consumer invalid offset '%s', using 'oldest'",
+			k.Offset)
 		return 0, err
 	}
+	consumerGroup, err := sarama.NewConsumerGroupFromClient(k.ConsumerGroup, newClient)
+	if err != nil {
+		logger.Warningf(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM",
+			"failed to creating consumer group client, [group]%s", k.ConsumerGroup)
+		return 0, err
+	}
+	k.consumerGroupClient = consumerGroup
+	cancelCtx, cancel := ctx.WithCancel(k.context.GetRuntimeContext())
+	k.cancelConsumer = cancel
+	k.wg = &sync.WaitGroup{}
+	k.wg.Add(1)
+	go func() {
+		defer k.wg.Done()
+		for {
+			if err := k.consumerGroupClient.Consume(cancelCtx, k.Topics, k); err != nil {
+				logger.Error(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM", "Error from kafka consumer", err)
+				return
+			}
+			// check if context was canceled, signaling that the consumer should stop
+			if cancelCtx.Err() != nil {
+				logger.Info(k.context.GetRuntimeContext(), "Consumer was canceled. Leaving consumer group")
+				return
+			}
+			k.ready = make(chan bool)
+		}
+	}()
+	<-k.ready
 	return 0, nil
 }
 
@@ -121,84 +179,87 @@ func (k *InputKafka) Description() string {
 }
 
 func (k *InputKafka) Start(collector pipeline.Collector) error {
-	k.shutdown = make(chan struct{})
-
-	// consume errors
+	k.collectorV1 = collector
+	k.version = v1
 	go func() {
-		for err := range k.cluster.Errors() {
-			select {
-			case <-k.shutdown:
-				logger.Infof(k.context.GetRuntimeContext(), "shutdown errors func")
-				return
-			default:
-				logger.Errorf(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM", "Error: %s\n", err.Error())
-			}
-		}
-		logger.Infof(k.context.GetRuntimeContext(), "shutdown errors loop")
+		k.receiver()
 	}()
-
-	// consume notifications
-	go func() {
-		for ntf := range k.cluster.Notifications() {
-			select {
-			case <-k.shutdown:
-				logger.Infof(k.context.GetRuntimeContext(), "shutdown notifications func")
-				return
-			default:
-				logger.Infof(k.context.GetRuntimeContext(), "Rebalanced: %v", ntf)
-			}
-		}
-		logger.Infof(k.context.GetRuntimeContext(), "shutdown notifications loop")
-	}()
-	k.wg = &sync.WaitGroup{}
-	k.wg.Add(1)
-	go func() {
-		defer k.wg.Done()
-		k.receiver(collector)
-	}()
-
-	logger.Infof(k.context.GetRuntimeContext(), "Started the kafka consumer service, brokers: %v, topics: %v",
-		k.Brokers, k.Topics)
 	return nil
 }
 
-func (k *InputKafka) receiver(collector pipeline.Collector) {
+func (k *InputKafka) StartService(context pipeline.PipelineContext) error {
+	k.collectorV2 = context.Collector()
+	k.version = v2
+	go func() {
+		k.receiver()
+	}()
+	return nil
+}
+
+func (k *InputKafka) receiver() {
+	for msg := range k.messages {
+		k.onMessage(msg)
+	}
+}
+
+// Setup implements ConsumerGroupHandler
+func (k *InputKafka) Setup(session sarama.ConsumerGroupSession) error {
+	k.readyCloser.Do(func() {
+		close(k.ready)
+	})
+	return nil
+}
+
+// Cleanup implements ConsumerGroupHandler
+func (k *InputKafka) Cleanup(sarama.ConsumerGroupSession) error {
+	return nil
+}
+
+// ConsumeClaim implements ConsumerGroupHandler, must start a consumer loop of ConsumerGroupClaim's Messages().
+func (k *InputKafka) ConsumeClaim(session sarama.ConsumerGroupSession, claim sarama.ConsumerGroupClaim) error {
+	logger.Debug(k.context.GetRuntimeContext(), "Consuming messages [partition]", claim.Partition(), "[topic]", claim.Topic(),
+		"init [offset]", claim.InitialOffset())
 	for {
 		select {
-		case msg, ok := <-k.cluster.Messages():
-			if ok {
-				k.onMessage(collector, msg)
-				k.markOffset(msg) // mark message as processed
+		case msg, ok := <-claim.Messages():
+			if !ok {
+				return nil
 			}
-		case <-k.shutdown:
-			logger.Infof(k.context.GetRuntimeContext(), "shutdown receiver func")
-			return
+			k.messages <- msg
+			session.MarkMessage(msg, "")
+		case <-session.Context().Done():
+			logger.Debug(k.context.GetRuntimeContext(), "Ctx was canceled, stopping consumerGroup")
+			return nil
 		}
 	}
 }
 
-func (k *InputKafka) markOffset(msg *sarama.ConsumerMessage) {
-	k.cluster.MarkOffset(msg, "")
-}
-
-func (k *InputKafka) onMessage(collector pipeline.Collector, msg *sarama.ConsumerMessage) {
-	if len(msg.Value) > k.MaxMessageLen {
-		logger.Errorf(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM", "Message longer than max_message_len (%d > %d)",
-			len(msg.Value), k.MaxMessageLen)
-		return
-	}
-
+func (k *InputKafka) onMessage(msg *sarama.ConsumerMessage) {
 	if msg != nil {
-		fields := make(map[string]string)
-		fields[string(msg.Key)] = string(msg.Value)
-		collector.AddData(nil, fields)
+		switch k.version {
+		case v1:
+			fields := make(map[string]string)
+			fields[string(msg.Key)] = string(msg.Value)
+			k.collectorV1.AddData(nil, fields)
+		case v2:
+			data, err := k.decoder.DecodeV2(msg.Value, nil)
+			if err != nil {
+				logger.Warning(k.context.GetRuntimeContext(), "DECODE_MESSAGE_FAIL_ALARM", "decode message failed", err)
+				return
+			}
+			k.collectorV2.CollectList(data...)
+		}
 	}
 }
 
 func (k *InputKafka) Stop() error {
-	close(k.shutdown)
+	k.readyCloser.Do(func() {
+		close(k.ready)
+	})
 	k.wg.Wait()
-	if err := k.cluster.Close(); err != nil {
+	k.cancelConsumer()
+	err := k.consumerGroupClient.Close()
+	if err != nil {
 		e := fmt.Errorf("[inputs.kafka_consumer] Error closing consumer: %v", err)
 		logger.Errorf(k.context.GetRuntimeContext(), "INPUT_KAFKA_ALARM", "%v", e)
 		return e
@@ -217,11 +278,13 @@ func init() {
 			ClientID:      "",
 			Topics:        nil,
 			Brokers:       nil,
-			MaxMessageLen: maxInputKafkaLen,
 			Version:       "",
 			Offset:        "oldest",
 			SASLUsername:  "",
 			SASLPassword:  "",
+			Assignor:      "range",
+			messages:      make(chan *sarama.ConsumerMessage, 256),
+			ready:         make(chan bool),
 		}
 	}
 }